Use an AtomicUint for tracking logging verbosity
This is safer than using a regular uint, as it eliminates
the chance of undefined data races by reading/writing to the
verbosity concurrently.

Since we don't care too much about the precise ordering of load/store
instructions here, we can use the Relaxed ordering but still get
atomic operations.